+++ leafletmap.php
@@ -1,353 +1,367 @@
 <?php
 
 /**
- * @plugin         Leaflet Map
- * @version         1.0.5
- * 
- * @author          amaral <grant@lrio.com
- * @link            http://lrio.com
- * @copyright       Copyright © 2023 Amaral All Rights Reserved
- * @license         http://www.gnu.org/licenses/gpl-2.0.html GNU/GPL
+ * @plugin   Leaflet Map
+ * @version  1.0.7
+ * @author   Grant Amaral <grant@lrio.com>
+ * @link     https://lrio.com
+ * @license  GPL-2.0-or-later
  */
+
 defined('_JEXEC') or die;
+
+use Joomla\CMS\Factory;
 use Joomla\CMS\Plugin\CMSPlugin;
-use Joomla\CMS\Language\Text;
+use Joomla\CMS\Uri\Uri;
 
-class PlgSystemLeafletMap extends CMSPlugin
-    {
-        // ga july 23 load the language files
-        protected $autoloadLanguage = true;
+class PlgSystemLeafletMap extends CMSPlugin {
+    protected $autoloadLanguage = true;
 
-        // Track initialized map containers
-        private $initializedMapContainers = [];
+    /** Ensure we only register assets once per request */
+    private bool $assetsRegistered = false;
 
-        // Counter for auto-incrementing container IDs
-        private $mapContainerCounter = 0;
+    /** Auto-increment container IDs */
+    private int $mapContainerCounter = 0;
 
-        public function onContentPrepare($context, &$article, &$params, $limitstart) {
-                // Check if the article content contains the plugin tag
-                if (strpos($article->text, '{leafletmap') !== false) {
-                    // Load the necessary CSS and JS files
-                    $this->loadAssets();
-
-                    // Replace the plugin tag with the generated map HTML
-                    $article->text = preg_replace_callback('/{leafletmap(.*?)}/is', array($this, 'replaceMapTag'), $article->text);
-                }
-            }
-
-        private function loadAssets() // Load Leaflet CSS and JS files
-            {
-                
-                JFactory::getDocument()->addStyleSheet('/plugins/system/leafletmap/css/leaflet.css');
-                JFactory::getDocument()->addStyleSheet('/plugins/system/leafletmap/css/Control.FullScreen.css');
-                JFactory::getDocument()->addStyleSheet('/plugins/system/leafletmap/css/leaflet_custom.css');
-                JFactory::getDocument()->addScript('/plugins/system/leafletmap/js/leaflet.js');
-                JFactory::getDocument()->addScript('//api.tiles.mapbox.com/mapbox.js/plugins/leaflet-omnivore/v0.3.1/leaflet-omnivore.min.js');
-                JFactory::getDocument()->addScript('/plugins/system/leafletmap/js/Control.FullScreen.js');
-            }        
-        private function replaceMapTag($matches) {
-            $params = $this->parseParams($matches[1]);
-                // Get the plugin parameters from the backend
-                $lat = $this->params->get('latitude', '');
-                $long = $this->params->get('longitude', '');
-                $markerTitle = $this->params->get('marker_title', '');
-                $width = $this->params->get('div_width', '');
-                $height = $this->params->get('div_height', '');
-                $basemap = $this->params->get('basemap', '');
-                $zoom = $this->params->get('zoom', '');
-                $kml_toggle = $this->params->get('kml_toggle', '');
-                $kml = $this->params->get('kml', '');
-                $geojson = $this->params->get('geojson', '');
-                $title_toggle = $this->params->get('title_toggle', '');
-                $zoom_wheel = $this->params->get('zoom_wheel', '');
-                $marker_icon = $this->params->get('marker_icon', '');
-                $toggle_css = $this->params->get('toggle_css', 0);
-            
-            // Override the parameters with values from the content call, if available
-                if (isset($params['lat'])) {
-                    $lat = $params['lat'];
-                }
-                if (isset($params['long'])) {
-                    $long = $params['long'];
-                }
-                if (isset($params['marker_title'])) {
-                    $markerTitle = $params['marker_title'];
-                }
-                if (isset($params['width'])) {
-                    $width = $params['width'];
-                }
-                if (isset($params['height'])) {
-                    $height = $params['height'];
-                }
-                if (isset($params['basemap'])) {
-                    $basemap = $params['basemap'];
-                }
-                if (isset($params['zoom'])) {
-                    $zoom = $params['zoom'];
-                }
-                if (isset($params['kml'])) {
-                    $kml = $params['kml'];
-                }
-                if (isset($params['geojson'])) {
-                    $geojson = $params['geojson'];
-                }
-                if (isset($params['title_toggle'])) {
-                    $title_toggle = $params['title_toggle'];
-                }
-                if (isset($params['zoom_wheel'])) {
-                    $zoom_wheel = $params['zoom_wheel'];
-                }
-
-            // Generate the map container ID with auto-increment
-            $mapContainerId = 'map' . $this->mapContainerCounter;
-                $this->mapContainerCounter++;
-                // Check if the map container has already been initialized
-                if (in_array($mapContainerId, $this->initializedMapContainers)) {
-                    return '<div id="' . $mapContainerId . '"></div>';
-                    }
-
-                // Set the flag to indicate the map container has been initialized
-                $this->initializedMapContainers[] = $mapContainerId;
-        $float = 'right';
-        if ($float == 1) {
-            $float_class = "float-left ";
-        } 
-        if ($float == 'right') {
-            $float_class = "float-right ";
-        } else {
-            $float_class ="";
+    public function onContentPrepare($context, &$article, &$params, $limitstart) {
+        if (strpos($article->text ?? '', '{leafletmap') === false) {
+            return;
         }
-        var_dump($float_class);
-                // Get the map container options from the content call or use defaults
-                    // $mapContainerOptions = $this->getContainerOptions($params, 'map');
-                    $mapContainerOptions = $this->getContainerOptions($params, ' . $mapContainerId . ');
-                    // $mapContainerOptions['style'] = 'width: ' . $width . '; height: ' . $height . '; float: ' . $float_style . ';';
-                    $mapContainerOptions['style'] = 'width: ' . $width . '; height: ' . $height . '; ';
-                    $mapContainerOptions['class'] = $float_class;
 
-                    $markerPopupCode = 'L.marker([' . $lat . ', ' . $long . ']).addTo(map_' . $mapContainerId . ').bindPopup(\'' . Text::_($markerTitle) . '\')';
-                if ($title_toggle == 0) {
-                        $markerPopupCode .= '.openPopup()';
-                    }
-                    $markerPopupCode .= ';';
-            // BaseLayer logic
-                if ($basemap == 0) {
-                    $basemap = "osm"; // default
-                } elseif ($basemap == 1) {
-                    $basemap = 'esriSat'; 
-                } elseif ($basemap == 2) {
-                    $basemap = 'esriTopo'; 
-                } elseif ($basemap == 3) {
-                    $basemap = 'openTopo';
-                }
-            // hide the flag 
-                if ($toggle_css == 1) {
-                    JFactory::getDocument()->addStyleSheet('/plugins/system/leafletmap/css/leaflet_flag.css');
-                    }
-               
-            // Generate the map HTML with the specified container options and initialization script
-                $html = '<div id="' . $mapContainerId . '" ' . $this->renderHtmlAttributes($mapContainerOptions) . '></div>';
-                $html .= '<script>
-                    document.addEventListener(\'DOMContentLoaded\', function() {
-                        var mapContainer_' . $mapContainerId . ' = document.getElementById(\'' . $mapContainerId . '\');
+        // Load assets once
+        $this->loadAssets();
 
-                        // basemaps
-                            const osm = L.tileLayer("https://tile.openstreetmap.org/{z}/{x}/{y}.png", {maxZoom: 19, attribution: "© OpenStreetMap" });
-                            const openTopo = L.tileLayer("https://{s}.tile.opentopomap.org/{z}/{x}/{y}.png", {maxZoom: 19, attribution: "@OpenTopoMap" });
-                            const esriSat = L.tileLayer("https://server.arcgisonline.com/ArcGIS/rest/services/World_Imagery/MapServer/tile/{z}/{y}/{x}", {attribution: "&copy; Esri" });
-                            const esriTopo = L.tileLayer("https://server.arcgisonline.com/ArcGIS/rest/services/World_Topo_Map/MapServer/tile/{z}/{y}/{x}", {attribution: "&copy; Esri" });
-
-                        var greenIcon = new L.Icon({
-                            iconUrl: "https://raw.githubusercontent.com/pointhi/leaflet-color-markers/master/img/marker-icon-green.png",
-                            shadowUrl: "https://cdnjs.cloudflare.com/ajax/libs/leaflet/0.7.7/images/marker-shadow.png",
-                            iconSize: [25, 41],
-                            iconAnchor: [12, 41],
-                            popupAnchor: [1, -34],
-                            shadowSize: [41, 41]
-                            });
-                            greenIcon = "";
-
-                        
-                        var map_' . $mapContainerId . ';
-                    
-                        if (' . ($zoom_wheel ? 'false' : 'true') . ') {
-                            map_' . $mapContainerId . ' = L.map(mapContainer_' . $mapContainerId . ', { dragging: !L.Browser.mobile, tap: !L.Browser.mobile, layers: ' . $basemap . ' }).setView([' . $lat . ', ' . $long . '], ' . $zoom . ');
-                            } else {
-                                map_' . $mapContainerId . ' = L.map(mapContainer_' . $mapContainerId . ', { scrollWheelZoom: false, dragging: !L.Browser.mobile, tap: !L.Browser.mobile, layers: ' . $basemap . ' }).setView([' . $lat . ', ' . $long . '], ' . $zoom . ');
-                            }
-
-                        var marker_' . $mapContainerId . ' = L.marker([' . $lat . ', ' . $long . ']).addTo(map_' . $mapContainerId . ').bindPopup(\'' . Text::_($markerTitle) . '\');';
-                        // var marker_' . $mapContainerId . ' = L.marker([' . $lat . ', ' . $long . '], {icon: greenIcon}).addTo(map_' . $mapContainerId . ').bindPopup(\'' . Text::_($markerTitle) . '\');';
-
-                        if ($title_toggle == 1) {
-                            $html .= 'marker_' . $mapContainerId . '.openPopup();';
-                        }
-
-                    $html .= '
-                        var baseMaps = {
-                            "OSM": osm,
-                            "Satellite": esriSat,
-                            "Topo": esriTopo,
-                            "Open Topo": openTopo
-                        };
-
-                        var layerControl = L.control.layers(baseMaps).addTo(map_' . $mapContainerId . ');
-
-                        // this turns off the leaflet atttribution. ukrane flag issue
-                        //map_' . $mapContainerId . '.attributionControl.setPrefix("");
-                        
-                        // geojson
-                            var geojsonLayer = null;
-                            var bounds = null;
-                            var geojsonUrl = "' . $geojson . '";
-
-                            if (true && geojsonUrl !== "") {
-                                geojsonLayer = L.geoJSON(null, {
-                                    pointToLayer: function(feature, latlng) {
-                                        // Use the green icon for markers in the KML layer
-                                        // return L.marker(latlng, { icon: greenIcon });
-                                        return L.marker(latlng);
-                                    },
-                                    onEachFeature: function(feature, layer) {
-                                        if (layer instanceof L.Marker) {
-                                            layer.bindPopup(feature.properties.name);
-                                        }
-                                    }
-                                });
-
-                                fetch(geojsonUrl)
-                                    .then(response => response.json())
-                                    .then(data => {
-                                        geojsonLayer.addData(data);
-                                        bounds = geojsonLayer.getBounds();
-                                        map_' . $mapContainerId . '.fitBounds(bounds);
-                                    })
-                                    .catch(error => console.error(\'Error fetching GeoJSON:\', error));
-
-                                geojsonLayer.addTo(map_' . $mapContainerId . ');
-                            }
-
-                        // kml
-                        var kmlLayer = null;
-                        var bounds = null;
-                        var kmlUrl = "' . $kml . '";
-                        if (true && kmlUrl !== "") {
-                            kmlLayer = omnivore.kml(kmlUrl, null, L.geoJson(null, {
-                                pointToLayer: function(feature, latlng) {
-                                // Use the green icon for markers in the KML layer
-                                // return L.marker(latlng, { icon: greenIcon });
-                                return L.marker(latlng);
-                                },
-                                onEachFeature: function(feature, layer) {
-                                    if (layer instanceof L.Marker) {
-                                        layer.bindPopup(feature.properties.name);
-                                    }
-                                }
-                            }));
-
-                            kmlLayer.on(\'ready\', function() {
-                                bounds = kmlLayer.getBounds();
-                                map_' . $mapContainerId . '.fitBounds(bounds);
-                            });
-
-                            kmlLayer.addTo(map_' . $mapContainerId . ');
-                        }
-
-                        // create a fullscreen button and add it to the map
-                        L.control.fullscreen({
-                            position: "topleft",
-                            title: "Show me the fullscreen!",
-                            titleCancel: "Exit fullscreen mode",
-                            content: null,
-                            forceSeparateButton: true,
-                            forcePseudoFullscreen: true,
-                            fullscreenElement: false
-                        }).addTo(map_' . $mapContainerId . ');
-
-                        map_' . $mapContainerId . '.on(\'enterFullscreen\', function(){
-                            if (kmlLayer && bounds) {
-                                map_' . $mapContainerId . '.fitBounds(bounds);
-                            }
-                        });
-
-                        map_' . $mapContainerId . '.on(\'exitFullscreen\', function(){
-                            if (kmlLayer && bounds) {
-                                map_' . $mapContainerId . '.fitBounds(bounds);
-                            }
-                        });
-
-                    });
-                </script>';
-
-                    return $html;
-            }
-        private function getContainerOptions($params, $type)
-            {
-                $options = [];
-
-                // Check if container options are specified in the content call
-                if (isset($params[$type . '_container'])) {
-                    $containerParams = $this->parseParams($params[$type . '_container']);
-                    foreach ($containerParams as $key => $value) {
-                        $options[$key] = $value;
-                    }
-                }
-
-                return $options;
-            }
-
-
-        private function getUniqueContainerId($containerOptions)
-            {
-                $defaultContainerId = 'leaflet-map';
-
-                // If the default container id does not exist, return it
-                if (!in_array('id', array_keys($containerOptions))) {
-                    return $defaultContainerId;
-                }
-
-                // Check if the default container id exists in all articles on the page
-                $existingIds = [];
-
-                // Get all articles on the page
-                $articles = JModelLegacy::getInstance('Articles', 'ContentModel')->getItems();
-
-                // Loop through each article to collect existing container IDs
-                foreach ($articles as $article) {
-                    preg_match_all('/id=[\'"](.*?)[\'"]/i', $article->text, $matches);
-                    if (!empty($matches[1])) {
-                        $existingIds = array_merge($existingIds, $matches[1]);
-                    }
-                }
-
-                // Append a number to the default container id to make it unique
-                $counter = 1;
-                $uniqueId = $defaultContainerId;
-                while (in_array($uniqueId, $existingIds)) {
-                    $uniqueId = $defaultContainerId . '-' . $counter;
-                    $counter++;
-                }
-
-                return $uniqueId;
-            }
-
-        private function renderHtmlAttributes($attributes)
-            {
-                $output = '';
-                foreach ($attributes as $key => $value) {
-                    $output .= $key . '="' . $value . '" ';
-                }
-                return trim($output);
-            }
-                private function parseParams($paramsString)
-            {
-                $params = array();
-                preg_match_all('/(\w+)=(["\'])(.*?)\2/', $paramsString, $matches, PREG_SET_ORDER);
-                foreach ($matches as $match) {
-                    $key = $match[1];
-                    $value = $match[3];
-                    $params[$key] = $value;
-                }
-                return $params;
-            }
+        // Replace each {leafletmap ...} tag
+        $article->text = preg_replace_callback('/{leafletmap(.*?)}/is', [$this, 'replaceMapTag'], $article->text);
     }
+
+    private function loadAssets(): void {
+        if ($this->assetsRegistered) {
+            return;
+        }
+
+        $doc = \Joomla\CMS\Factory::getDocument();
+        $wa  = $doc->getWebAssetManager();
+
+        // CSS
+        $wa->registerAndUseStyle('leaflet-css',        'media/plg_system_leafletmap/css/leaflet.css',            [], ['version' => 'auto']);
+        $wa->registerAndUseStyle('leaflet-fs-css',     'media/plg_system_leafletmap/css/Control.FullScreen.css', [], ['version' => 'auto']);
+        $wa->registerAndUseStyle('leaflet-custom-css', 'media/plg_system_leafletmap/css/leaflet_custom.css',     [], ['version' => 'auto']);
+
+        // JS (order & deps matter)
+        $wa->registerAndUseScript('leaflet-js',         'media/plg_system_leafletmap/js/leaflet.js',               [],                 ['defer' => true, 'version' => 'auto']);
+        $wa->registerAndUseScript('leaflet-omni-js',    'media/plg_system_leafletmap/js/leaflet-omnivore.min.js',  ['leaflet-js'],     ['defer' => true, 'version' => 'auto']);
+        $wa->registerAndUseScript('leaflet-fs-js',      'media/plg_system_leafletmap/js/Control.FullScreen.js',    ['leaflet-js'],     ['defer' => true, 'version' => 'auto']);
+        $wa->registerAndUseScript('leaflet-helpers-js', 'media/plg_system_leafletmap/js/leaflet-helpers.js',       ['leaflet-js', 'leaflet-omni-js'], ['defer' => true, 'version' => 'auto']);
+
+        $this->assetsRegistered = true;
+    }
+
+
+    private function replaceMapTag(array $matches): string {
+        $params = $this->parseParams($matches[1]);
+
+        // Start with plugin defaults
+        $config = [
+            'lat'            => $this->params->get('lat'),
+            'lng'            => $this->params->get('lng'),
+            'longitude'      => $this->params->get('longitude'), // kept if used in your XML
+            'zoom'           => $this->params->get('zoom', 13),
+            'kml_toggle'     => (bool) $this->params->get('kml_toggle', 0),
+            'title_toggle'   => (bool) $this->params->get('title_toggle', 0),
+            'zoom_wheel'     => (bool) $this->params->get('zoom_wheel', 1),
+            'toggle_css'     => (bool) $this->params->get('toggle_css', 0),
+            'marker_title'   => (string) $this->params->get('marker_title', ''),
+            'marker_content' => (string) $this->params->get('marker_content', ''),
+            'marker_icon'    => (string) $this->params->get('marker_icon', ''),
+            'marker_custom'  => (string) $this->params->get('marker_custom', ''),
+            'basemap'        => (string) $this->params->get('basemap', ''),
+            'kml'            => (string) $this->params->get('kml', ''),
+            'geojson'        => (string) $this->params->get('geojson', ''),
+            'width'          => (string) $this->params->get('width', '100%'),
+            'height'         => (string) $this->params->get('height', '300px'),
+            'align'          => (string) $this->params->get('align', 'align-center'),
+            'markers'        => '' // shortcode-only composite param
+        ];
+
+        // Override with shortcode params
+        foreach ($params as $key => $value) {
+            if (array_key_exists($key, $config)) {
+                $config[$key] = $value;
+            }
+        }
+
+        // Normalize
+        $lat            = (float) $config['lat'];
+        $lng            = (float) $config['lng'];
+        $zoom           = (int)   $config['zoom'];
+        $kml_toggle     = (bool)  $config['kml_toggle'];
+        $title_toggle   = (bool)  $config['title_toggle'];
+        $zoom_wheel     = (bool)  $config['zoom_wheel'];
+        $toggle_css     = (bool)  $config['toggle_css'];
+        $markerTitle    = trim((string) $config['marker_title']);
+        $markerContent  = trim((string) $config['marker_content']);
+        $marker_icon    = trim((string) $config['marker_icon']);
+        $marker_custom  = trim((string) $config['marker_custom']);
+        $kml            = trim((string) $config['kml']);
+        $geojson        = trim((string) $config['geojson']);
+        $width          = trim((string) $config['width']);
+        $height         = trim((string) $config['height']);
+        $align          = strtolower(trim((string) $config['align']));
+
+        // Multiple markers from shortcode (lat,lng,title,content,icon | ...)
+        $markers = $this->parseMultipleMarkers($params['markers'] ?? '');
+
+        // Fallback to single marker if no KML/GeoJSON and no multi-markers provided
+        if (empty($markers) && !$kml && !$geojson) {
+            $markers[] = [
+                'lat'     => $lat,
+                'lng'     => $lng,
+                'title'   => $markerTitle,
+                'content' => $markerContent,
+                'icon'    => $marker_icon
+            ];
+        }
+
+        // Validate dimensions
+        if (!preg_match('/^\d+(px|%)$/', $width)) {
+            $width  = '100%';
+        }
+        if (!preg_match('/^\d+(px|%)$/', $height)) {
+            $height = '300px';
+        }
+
+        // Validate alignment
+        $validAligns = ['align-left', 'align-center', 'align-right'];
+        if (!in_array($align, $validAligns, true)) {
+            $align = 'align-center';
+        }
+
+        // Marker icon handling
+        if ($marker_icon === 'custom' && $marker_custom !== '') {
+            $marker_icon = trim($marker_custom);
+        }
+
+        if ($marker_icon !== '') {
+            $isAbsolute =
+                str_starts_with($marker_icon, 'http://') ||
+                str_starts_with($marker_icon, 'https://') ||
+                str_starts_with($marker_icon, '//') ||
+                str_starts_with($marker_icon, '/') ||
+                str_starts_with($marker_icon, 'data:');
+
+            if (!$isAbsolute) {
+                // Bare filename like "blue-marker.png"
+                $marker_icon = Uri::root() . 'media/plg_system_leafletmap/images/markers/' . ltrim($marker_icon, '/');
+            }
+        }
+
+
+        // Hide the little attribution flag if requested
+        if ($toggle_css) {
+            Factory::getDocument()->addStyleDeclaration('.leaflet-attribution-flag{display:none !important;}');
+        }
+
+        // Escape popup content for JS
+        $markerContentJs = $this->escapeForJs($markerContent);
+
+        // Unique container
+        $mapContainerId = 'map' . $this->mapContainerCounter++;
+        $mapContainerOptions = $this->getContainerOptions($params, 'map');
+        $mapContainerOptions['style'] = 'width:' . $width . ';height:' . $height . ';';
+        $mapContainerOptions['class'] = $align;
+
+        // Build HTML
+        $html  = '<div class="map-container"><div id="' . $mapContainerId . '" ' . $this->renderHtmlAttributes($mapContainerOptions) . '></div></div>';
+
+        // Build JS
+        $markersJson    = $this->markersToJSON($markers);
+        $titleToggleJs  = $title_toggle ? 'true' : 'false';
+        $hasKmlJs       = $kml   !== '' ? 'true' : 'false';
+        $hasGeoJsonJs   = $geojson !== '' ? 'true' : 'false';
+        $kmlUrlJs       = $this->escapeForJs($kml);
+        $geoJsonUrlJs   = $this->escapeForJs($geojson);
+        $markerTitleJs  = $this->escapeForJs($markerTitle);
+        $markerIconJs   = $this->escapeForJs($marker_icon);
+        // add these two:
+        $baseMarkerDirJs = $this->escapeForJs(\Joomla\CMS\Uri\Uri::root() . 'media/plg_system_leafletmap/images/markers/');
+        $shadowUrlJs     = $this->escapeForJs(\Joomla\CMS\Uri\Uri::root() . 'media/plg_system_leafletmap/images/marker-shadow.png');
+
+        $showDefault       = (int) $this->params->get('show_default_marker', 1);
+        $hasShortcodeMarks = !empty($markers);
+        $hasKml            = ($kml !== '');
+        $hasGeojson        = ($geojson !== '');
+        $shouldAddDefault  = ($showDefault && !$hasShortcodeMarks && !$hasKml && !$hasGeojson);
+        // BEFORE the heredoc 1.0.8
+        $shouldAddDefaultJs = $shouldAddDefault ? 'true' : 'false';
+
+        // Ensure plugin-level icon is a full URL if it’s a filename
+        $resolvedIcon = trim($marker_icon);
+        if ($resolvedIcon === 'custom' && !empty($config['marker_custom'])) {
+            $resolvedIcon = trim($config['marker_custom']);
+        }
+        if (
+            $resolvedIcon !== '' &&
+            !preg_match('#^(?:https?:)?//#', $resolvedIcon) &&
+            $resolvedIcon[0] !== '/' &&
+            strpos($resolvedIcon, 'data:') !== 0
+        ) {
+            $resolvedIcon = \Joomla\CMS\Uri\Uri::root() . 'media/plg_system_leafletmap/images/markers/' . ltrim($resolvedIcon, '/');
+        }
+
+        $markerIconJs   = $this->escapeForJs($resolvedIcon);
+        $markerTitleJs  = $this->escapeForJs($markerTitle);
+        $markerContentJs = $this->escapeForJs($markerContent);
+        $shadowUrlJs    = $this->escapeForJs(\Joomla\CMS\Uri\Uri::root() . 'media/plg_system_leafletmap/images/marker-shadow.png');
+
+        $html .= <<<SCRIPT
+<script>
+document.addEventListener('DOMContentLoaded', function() {
+    if (typeof initLeafletMap === 'undefined') {
+        console.error('Leaflet helpers not loaded. Check media asset paths.');
+        return;
+    }
+
+    const map = initLeafletMap("{$mapContainerId}", {$lat}, {$lng}, {$zoom}, {$zoom_wheel});
+
+    // ---- Per-marker icon helpers ----
+    const BASE_MARKER_DIR = "{$baseMarkerDirJs}";
+    const SHADOW_URL      = "{$shadowUrlJs}";
+    const GLOBAL_ICON     = "{$markerIconJs}"; // plugin-level icon URL (may be empty)
+
+    function resolveIconUrl(raw) {
+        if (!raw) return "";
+        const r = String(raw).trim();
+        // absolute (http/https), protocol-relative (//), site-absolute (/), or data: URIs → use as-is
+        if (/^(?:https?:)?\\/\\//.test(r) || r.startsWith('/') || r.startsWith('data:')) return r;
+        // bare filename → prepend your media markers dir
+        return BASE_MARKER_DIR + r.replace(/^\\/+/, '');
+    }
+
+    // Multiple markers (supports per-marker icon via shortcode)
+    const markers = {$markersJson};
+    markers.forEach((m, i) => {
+        const opt = { title: m.title || '' };
+
+        // Prefer per-marker icon; fallback to plugin-level icon; else Leaflet default
+        const iconUrl = m.icon ? resolveIconUrl(m.icon) : (GLOBAL_ICON || "");
+        if (iconUrl) {
+            opt.icon = L.icon({
+                iconUrl,
+                iconSize: [25, 41],
+                iconAnchor: [12, 41],
+                popupAnchor: [1, -34],
+                shadowUrl: SHADOW_URL
+            });
+        }
+
+        const mrk = L.marker([m.lat, m.lng], opt).addTo(map);
+        if (m.content) mrk.bindPopup(m.content);
+        if (i === 0 && {$titleToggleJs} && m.content) mrk.openPopup();
+    });
+
+    // Fit bounds when multiple markers
+    if (markers.length > 1) {
+        const group = new L.featureGroup(markers.map(m => L.marker([m.lat, m.lng])));
+        map.fitBounds(group.getBounds().pad(0.1));
+    }
+
+    // Default marker only when the map would otherwise be empty (controlled by settings)
+    if (!markers.length && !{$hasKmlJs} && !{$hasGeoJsonJs} && {$shouldAddDefaultJs}) {
+        const opt = {};
+        if ("{$markerIconJs}") {
+            opt.icon = L.icon({
+                iconUrl: "{$markerIconJs}",
+                iconSize: [25, 41],
+                iconAnchor: [12, 41],
+                popupAnchor: [1, -34],
+                shadowUrl: "{$shadowUrlJs}"
+            });
+        }
+        const fallback = L.marker([{$lat}, {$lng}], opt).addTo(map);
+        // Prefer content; fall back to title
+        const popupHtml = "{$markerContentJs}" || "{$markerTitleJs}";
+        if (popupHtml) fallback.bindPopup(popupHtml);
+    }
+
+    // KML/GeoJSON (external helpers)
+    if ({$hasKmlJs})     { loadKml(map, "{$kmlUrlJs}", "Route Layer"); }
+    if ({$hasGeoJsonJs}) { loadGeoJson(map, "{$geoJsonUrlJs}", "Schedule Layer"); }
+});
+</script>
+SCRIPT;
+
+
+        return $html;
+    }
+
+    private function getContainerOptions(array $params, string $type): array {
+        $options = [];
+        if (isset($params[$type . '_container'])) {
+            $containerParams = $this->parseParams($params[$type . '_container']);
+            foreach ($containerParams as $key => $value) {
+                $options[$key] = $value;
+            }
+        }
+        return $options;
+    }
+
+    private function renderHtmlAttributes(array $attributes): string {
+        $out = '';
+        foreach ($attributes as $key => $value) {
+            $out .= htmlspecialchars($key, ENT_QUOTES) . '="' . htmlspecialchars((string) $value, ENT_QUOTES) . '" ';
+        }
+        return trim($out);
+    }
+
+    private function parseParams(string $paramsString): array {
+        $params = [];
+        preg_match_all('/(\w+)=(["\'])(.*?)\2/', $paramsString, $matches, PREG_SET_ORDER);
+        foreach ($matches as $m) {
+            $params[$m[1]] = $m[3];
+        }
+        return $params;
+    }
+
+    private function parseMultipleMarkers(string $markersString): array {
+        $markers = [];
+        if ($markersString === '') {
+            return $markers;
+        }
+        foreach (explode('|', $markersString) as $group) {
+            $parts = array_map('trim', explode(',', $group));
+            if (count($parts) >= 2) {
+                $markers[] = [
+                    'lat'     => (float) ($parts[0] ?? 0),
+                    'lng'     => (float) ($parts[1] ?? 0),
+                    'title'   => (string) ($parts[2] ?? ''),
+                    'content' => (string) ($parts[3] ?? ''),
+                    'icon'    => (string) ($parts[4] ?? '')
+                ];
+            }
+        }
+        return $markers;
+    }
+
+    private function markersToJSON(array $markers): string {
+        $processed = [];
+        foreach ($markers as $m) {
+            $content = (string) ($m['content'] ?? '');
+            $content = str_replace(["\r\n", "\r", "\n"], '<br>', $content);
+            $content = str_replace(['\\', '"'], ['\\\\', '\\"'], $content);
+            $processed[] = [
+                'lat'     => (float) ($m['lat'] ?? 0),
+                'lng'     => (float) ($m['lng'] ?? 0),
+                'title'   => (string) ($m['title'] ?? ''),
+                'content' => $content,
+                'icon'    => (string) ($m['icon'] ?? '')
+            ];
+        }
+        return json_encode($processed, JSON_UNESCAPED_SLASHES);
+    }
+
+    private function escapeForJs(string $s): string {
+        $s = str_replace(["\r\n", "\r", "\n"], ' ', $s);
+        return str_replace(['\\', '"'], ['\\\\', '\\"'], $s);
+    }
+}